requirements.txt
torch>=2.0.0
numpy>=1.21.0
PyYAML>=6.0
tensorboard>=2.11.0
matplotlib>=3.5.0
tqdm>=4.64.0
Pillow>=9.0.0
scikit-learn>=1.0.0

